+++ prmp/Spring Boot Backend/src/main/java/com/prmp/controller/UserController.java
@@ -28,20 +28,20 @@ public class UserController {
     }
 
     @GetMapping("/{id}")
-    public User getUserById(@PathVariable Integer id) {
+    public User getUserById(@PathVariable Long id) {
 
         return userService.getUserById(id);
     }
 
     @PutMapping("/{id}")
-    public User updateUser(@PathVariable Integer id,
+    public User updateUser(@PathVariable Long id,
                            @RequestBody User user) {
 
         return userService.updateUser(id, user);
     }
 
     @DeleteMapping("/{id}")
-    public String deleteUser(@PathVariable Integer id) {
+    public String deleteUser(@PathVariable Long id) {
 
         userService.deleteUser(id);
