+++ prmp/Spring Boot Backend/src/main/java/com/prmp/entity/User.java
@@ -20,7 +20,7 @@ public class User {
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Integer id;
+	private Long id;
 	
 	private String name;
 	
@@ -30,11 +30,11 @@ public class User {
 	@Enumerated(EnumType.STRING)
 	private UserRole role;
 
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
 
@@ -70,7 +70,7 @@ public class User {
 		this.role = role;
 	}
 
-	public User(Integer id, String name, String email, String password, UserRole role) {
+	public User(Long id, String name, String email, String password, UserRole role) {
 		
 		this.id = id;
 		this.name = name;
