+++ prmp/Spring Boot Backend/src/main/java/com/prmp/service/impl/UserServiceImpl.java
@@ -26,14 +26,14 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public User getUserById(Integer id) {
+    public User getUserById(Long id) {
 
         return userRepository.findById(id)
                 .orElseThrow(() -> new RuntimeException("User not found"));
     }
 
     @Override
-    public User updateUser(Integer id, User updatedUser) {
+    public User updateUser(Long id, User updatedUser) {
 
         User user = getUserById(id);
 
@@ -46,7 +46,7 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public void deleteUser(Integer id) {
+    public void deleteUser(Long id) {
         userRepository.deleteById(id);
     }
 }
